Move the static getContentUri methods to ProviderContract.Artwork
Position them alongside the other static methods, mirroring the separation between MuzeiArtProvider methods and their static equivalents in ProviderContract.Artwork.
