+++ android-client-common/src/main/java/com/google/android/apps/muzei/sync/ArtworkLoadWorker.kt
@@ -106,7 +106,7 @@ class ArtworkLoadWorker : Worker() {
         if (BuildConfig.DEBUG) {
             Log.d(TAG, "Artwork Load for $componentName")
         }
-        val contentUri = MuzeiArtProvider.getContentUri(applicationContext, componentName)
+        val contentUri = ProviderContract.Artwork.getContentUri(applicationContext, componentName)
         try {
             ContentProviderClientCompat.getClient(applicationContext, contentUri)?.use { client ->
                 val result = client.call(METHOD_GET_LOAD_INFO)
